tests/wpt/tests/html/browsers/the-window-object/accessing-other-browsing-contexts/iterator.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>window[@@iterator]</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_false(Symbol.iterator in window);
});
</script>
